+++ src/components/Form/Button.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
-import buttonBG from '../assets/hc_button.svg'
-import buttonBGHover from '../assets/hc_button_hover.svg'
-import { hexToRgba } from '../utility/utilities'
+import buttonBG from '../../assets/hc_button.svg'
+import buttonBGHover from '../../assets/hc_button_hover.svg'
+import { hexToRgba } from '../../utility/utilities'
 
 const buttonWidth = {
   small: '40px',
