+++ src/components/Form/index.js
@@ -1,0 +1,8 @@
+import Button from './Button'
+import Checkbox from './Checkbox'
+import Select from './Select'
+import TextArea from './TextArea'
+import TextInput from './TextInput'
+import ToggleSwitch from './ToggleSwitch'
+import Dropdown from './Dropdown'
+export { Button, Checkbox, Select, TextArea, TextInput, ToggleSwitch, Dropdown }
