+++ src/pages/Login.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react'
 import { ThemeContext } from 'styled-components'
 import styled from 'styled-components'
 import Landing from '../containers/Landing'
-import Button from '../components/Button'
+import { Button } from '../components/Form'
 import google from '../assets/icons/google.svg'
 import github from '../assets/icons/github.svg'
